#74 Use standard .NET TLS version instead of OS TLS version.
This prevents TLS (SChannel 70, EventID 36887) errors and hence unsuccessful connections on Windows versions pre Windows 10. For example, the standard TLS version of Windows 7 is TLS 1.0 which is deprecated.

The settings in this commit sets the TLS version to the standard TLS version of the currently installed .NET Framework (4.7.X) instead of the System/OS default TLS version.