Fix issue #81 Broken login due to changed privacy consent agreement
The old code cannot handle the login any more, when a new privacy consent agreement is requested by Tumblr, and falsely interprets it as a TFA request.

The change re-enables the old authentication dialog, a small browser window where the user can login with his credentials and accept the privacy consent agreement. As the previous browser component is no longer able to load modern pages/JS, a new one was necessary. MS WebView doesn't work with Admin accounts and/or disabled UAC. So CefSharp (Chromium Embedded Framework) is used.

This change
- has a notably higher payload due to the integration of CEF
- doesn't support AnyCPU builds any more (for now)
- is supposed to discontinue the support for OS below Win10
- currently only builds x64 builds

Committed changes:
- fixed project build configurations
- fixed AppVeyor build configuration and script